Revert "Release/w2"
