--- .env
@@ -1,8 +1,0 @@
-REACT_APP_GOOGLE_API_KEY=AIzaSyDB1Rb-qZEOK1TCZlNjtcSRxxJZa7vm3W8
-REACT_APP_GOOGLE_AUTHDOMAIN=upbeat-button-327904.firebaseapp.com
-REACT_APP_GOOGLE_PROJECTID=upbeat-button-327904
-REACT_APP_GOOGLE_STORAGEBUCKET=upbeat-button-327904.appspot.com
-REACT_APP_GOOGLE_MESSAGING_SENDERID=378435873818
-REACT_APP_GOOGLE_APPID=1:378435873818:web:fb66772e7154b0a0b7f5a3
-REACT_APP_GOOGLE_MEASUREMENTID=G-BEXQFLEJBV
-REACT_APP_API_SERVER=http://localhost:8080
