+++ package.json
@@ -10,7 +10,6 @@
     "@mui/material": "^5.0.1",
     "@mui/styled-engine-sc": "^5.0.0",
     "@mui/styles": "^5.0.1",
-    "@reduxjs/toolkit": "^1.6.2",
     "@sentry/react": "^6.13.2",
     "@sentry/tracing": "^6.13.2",
     "@testing-library/jest-dom": "^5.11.4",
@@ -21,14 +20,13 @@
     "dotenv": "^10.0.0",
     "firebase": "^9.1.1",
     "react": "^17.0.2",
-    "react-circular-progressbar": "^2.0.4",
     "react-dom": "^17.0.2",
     "react-google-login": "^5.2.2",
-    "react-redux": "^7.2.5",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
-    "react-tradingview-widget": "^1.3.2",
     "react-with-firebase-auth": "^1.3.11",
+    "redux": "^4.1.1",
+    "redux-thunk": "^2.3.0",
     "styled-components": "^5.3.1",
     "web-vitals": "^0.2.4",
     "workbox-background-sync": "^5.1.3",
