--- src/components/Challenge/styles.js
@@ -1,24 +1,0 @@
-import { makeStyles } from "@mui/styles";
-
-export const useStyles = makeStyles({
-  root: {},
-  challengeTitle: {
-    marginTop: "5.5rem",
-    marginBottom: ".25rem",
-  },
-  refreshButton: {
-    borderRadius: "12px",
-    padding: "6px",
-    minWidth: "auto",
-    float: "right",
-  },
-  updateDetails: {
-    float: "right",
-    fontSize: "14px",
-  },
-  updateDetailsText: {
-    display: "inline-block",
-    paddingRight: "10px",
-    lineHeight: "36px",
-  },
-});
