--- src/components/ChallengeList/ChallengeCard/ChallengeCard.jsx
@@ -1,41 +1,0 @@
-import React from "react";
-
-// material ui imports
-import Card from "@mui/material/Card";
-import { Typography } from "@mui/material";
-import { Button } from "@mui/material";
-
-import { useStyles } from "./styles";
-import { Link } from "react-router-dom";
-
-import { Timestamp } from 'firebase/firestore';
-
-export default function ChallengeCard(props) {
-  const classes = useStyles();
-  const { name, id, startDate, endDate } = props.challenge;
-
-
-  let s_date = new Timestamp(startDate._seconds, startDate._nanoseconds);
-  let e_date = new Timestamp(endDate._seconds, endDate._nanoseconds);
-
-  return (
-    <Card variant="outlined" className={classes.root}>
-      <div className={classes.challengeDetails}>
-        <Typography variant="h5">{name}</Typography>
-        <Typography variant="h6">
-          {s_date.toDate().toDateString()} - {e_date.toDate().toDateString()}
-        </Typography>
-      </div>
-      <Link className={classes.styledLink} to={`/challenge/${id}`}>
-        <Button
-          color="primary"
-          fullWidth="true"
-          style={{ justifyContent: "flex-start" }}
-          variant="contained"
-        >
-          Create Portfolio
-        </Button>
-      </Link>
-    </Card>
-  );
-}
