--- src/components/ChallengeList/ChallengeCard/styles.js
@@ -1,16 +1,0 @@
-import { makeStyles } from "@mui/styles";
-
-export const useStyles = makeStyles({
-  root: {
-    padding: "1.5rem 2rem",
-  },
-  challengeDetails: {
-    marginBottom: "2rem",
-  },
-  cardButton: {
-    justifyContent: "start",
-  },
-  styledLink: {
-    textDecoration: "none",
-  },
-});
