--- src/components/ChallengeList/styles.js
@@ -1,25 +1,0 @@
-import { makeStyles } from "@mui/styles";
-
-export const useStyles = makeStyles({
-  root: {
-    padding: "1.75rem",
-  },
-  sensexChartCard: {
-    height: "55vh",
-    padding: ".25rem",
-  },
-  challengeFilterCard: {
-    position: "sticky",
-    top: "5.5rem",
-    padding: "1rem",
-  },
-  challengeList: {
-    padding: "1rem",
-  },
-  challengeListTitle: {
-    padding: "0rem 1rem",
-    paddingTop: "1.5rem",
-    position: "sticky",
-    top: "5.5rem",
-  },
-});
