--- src/components/EmailLogin/EmailLoginStyle.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const EmailLoginTitle = styled.div`
-  font-weight: 600;
-  font-size: 24px;
-  line-height: 32px;
-  margin-bottom: 33px;
-`;
