--- src/components/Header/Header.jsx
@@ -1,74 +1,0 @@
-import React from "react";
-
-// mui components
-import AppBar from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import Avatar from "@mui/material/Avatar";
-import MenuItem from "@mui/material/MenuItem";
-import Menu from "@mui/material/Menu";
-import { makeStyles } from "@mui/styles";
-
-// custom components
-import Logo from "../Logo/logo";
-
-// redux store
-import { useSelector } from "react-redux";
-
-const useStyles = makeStyles({
-  profileIcon: {},
-  toolbar: {
-    display: "flex",
-    justifyContent: "space-between",
-  },
-});
-
-export default function Header() {
-  const classes = useStyles();
-  const user = useSelector((state) => state.user);
-  const [anchorEl, setAnchorEl] = React.useState(null);
-
-  const handleMenu = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  return (
-    <AppBar>
-      <Toolbar className={classes.toolbar}>
-        <Logo light />
-        <div>
-          <Avatar
-            alt={user.name}
-            src={user.profileImage}
-            aria-label="account of current user"
-            aria-controls="menu-appbar"
-            aria-haspopup="true"
-            onClick={handleMenu}
-            className={classes.profileIcon}
-          />
-          <Menu
-            id="menu-appbar"
-            anchorEl={anchorEl}
-            anchorOrigin={{
-              vertical: "bottom",
-              horizontal: "right",
-            }}
-            keepMounted
-            transformOrigin={{
-              vertical: "top",
-              horizontal: "right",
-            }}
-            open={Boolean(anchorEl)}
-            onClose={handleClose}
-          >
-            <MenuItem onClick={handleClose}>Profile</MenuItem>
-            <MenuItem onClick={handleClose}>Logout</MenuItem>
-          </Menu>
-        </div>
-      </Toolbar>
-    </AppBar>
-  );
-}
