+++ src/components/Home/Home.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Home() {
+  return <div>Home</div>;
+}
