+++ src/components/Logo/logo.js
@@ -1,21 +1,18 @@
 import React from 'react';
 import { withStyles } from '@mui/styles';
-import logoLongDark from '../../assets/logo-long-dark.svg';
-import logoLonglight from '../../assets/logo-long-light.svg';
+import logo from '../../assets/logo.svg';
 import PropTypes from 'prop-types';
 
 const styles = {
     root: {
+        padding: '1rem',
         maxWidth: '25rem'
     },
 };
 
 function Logo(props) {
-    const { classes, light } = props;
-    return (
-        <img className={classes.root} src={light ? logoLonglight : logoLongDark} alt="Fantasy Portfolio League" />
-        );
-
+    const { classes } = props;
+    return <img className={classes.root} src={logo} alt="Fantasy Portfolio League" />;
 }
 
 Logo.propTypes = {
