--- src/components/StockDetails/StockDetails.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function StockDetails() {
-  return (
-    <div>
-      StockDetails
-    </div>
-  )
-}
