+++ src/index.css
@@ -2,7 +2,7 @@
 
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont,'Inter', sans-serif !important;
+  font-family: -apple-system, BlinkMacSystemFont,'Inter', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background-image: url('./assets/background.png');
