+++ src/index.js
@@ -6,10 +6,6 @@ import "./index.css";
 import * as Sentry from '@sentry/react'
 import { Integrations } from '@sentry/tracing'
 
-// state store - redux 
-import store from './store';
-import { Provider } from 'react-redux';
-
 import App from "./App";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import reportWebVitals from "./reportWebVitals";
@@ -26,13 +22,11 @@ Sentry.init({
 })
 
 ReactDOM.render(
-  <Provider store={store}>
-    <React.StrictMode>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
-    </React.StrictMode>
-  </Provider>,
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>,
   document.getElementById("root")
 );
